Cast in linear scale. Find minmax now returns 0 and 255 for colour images rather than doing nothing.
git-svn-id: https://subversion.assembla.com/svn/ATD_Libraries/FreeImageAlgorithms@874 422d053a-e635-4566-91ac-90b073d95016

+++ trunk/src/FreeImageAlgorithms_LinearScale.cpp
@@ -209,13 +209,13 @@ template < class Tsrc > FIBITMAP * LINEAR_SCALE < Tsrc >::convert (FIBITMAP * sr
         {
             val = src_bits[x];
 
-            if (val <= min_found)
+            if ((double)val <= min_found)
             {
 
                 dst_bits[x] = 0;
                 continue;
             }
-            else if (val >= max_found)
+            else if ((double)val >= max_found)
             {
                 dst_bits[x] = 255;
             }
